Major Revision P1 Resource Collection
Only Command HQ has actual resource storage amounts over 1. This is to make the Maximum more uniform across all Factions, and to give a player who has lost one key structure an ability to continue playing. HQ each stores 5000 credites for a max of 10000 total. 5000 alone is plenty to work with. added credit storage unit to store 10000 credits however not implemented in game. For a P3 sometime in the future :)